requirements.txt
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
python-dotenv==1.0.1
cryptography==42.0.5